pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dev-template"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
repository = "https://github.com/valory-xyz/dev-template"
keywords = [ "autonomy", "autonomous", "autonolas", "service", "multi", "agent",]
classifiers = [ "Environment :: Console", "Environment :: Web Environment", "Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Natural Language :: English", "Operating System :: MacOS", "Operating System :: Microsoft", "Operating System :: Unix", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Topic :: Communications", "Topic :: Internet", "Topic :: Scientific/Engineering", "Topic :: Software Development", "Topic :: System",]
[[tool.poetry.packages]]
include = "packages"

[tool.poetry.dependencies]
python = "^3.8"
open-autonomy = "==0.12.1.post4"
toml = "^0.10.2"
open-aea-ledger-ethereum = "1.39.0.post1"
hypothesis = "^6.90.0"
grpcio = "^1.59.3"
typing-extensions = "^4.8.0"
py-ecc = "^6.0.0"
asn1crypto = "^1.5.1"

[tool.poetry.group.dev.dependencies.tomte]
version = "==0.2.13"
extras = [ "cli", "tests",]
